requirements.txt
requests
python-kasa
flask
python-dotenv
